Version 3
HelpExtractor v2 exhibited odd behavior when one attempted to source a plugin using it from a non-standard location.  Remember, the HelpExtractor is trying to extract documentation from the file, put it in the .vim/doc (vimfiles\doc) directory, and remove the extractor code.  An error message is now produced and the extraction is "finish"d, leaving the plugin operational but with the extractor code and help intact (but not available via :help).